+++ changelog.md
@@ -7,6 +7,10 @@
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot)</br>
 
+### Changed
+
+- command buttons automatically disable after they expire
+
 ## [4.1.4] - 2025-03-29
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot/commit/b1dfdebedf21d46011275dfe4054d96d3cf0d690)</br>